requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cloudinary==1.29.0
cryptography==37.0.4
defusedxml==0.7.1
dj-database-url==0.5.0
dj-email-url==1.0.5
Django==4.0.6
django-allauth==0.51.0
django-cache-url==3.4.2
django-cloudinary-storage==0.3.0
django-livereload-server==0.4
django-shopping-cart==0.1
environs==9.5.0
gunicorn==20.1.0
idna==3.3
marshmallow==3.17.0
oauthlib==3.2.0
packaging==21.3
Pillow==9.2.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dotenv==0.20.0
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tornado==6.2
tzdata==2022.1
urllib3==1.26.10
whitenoise==6.2.0
